+++ lib/base/src/main/java/com/blankj/base/mvp/BasePresenter.java
@@ -1,15 +1,16 @@
 package com.blankj.base.mvp;
 
+import android.app.Activity;
+
+import com.blankj.utilcode.util.ActivityUtils;
 import com.blankj.utilcode.util.LogUtils;
+import com.blankj.utilcode.util.Utils;
 
 import java.util.HashMap;
 import java.util.Map;
 
-import androidx.annotation.CallSuper;
+import androidx.annotation.NonNull;
 import androidx.lifecycle.Lifecycle;
-import androidx.lifecycle.LifecycleObserver;
-import androidx.lifecycle.LifecycleOwner;
-import androidx.lifecycle.OnLifecycleEvent;
 
 /**
  * <pre>
@@ -19,7 +20,7 @@ import androidx.lifecycle.OnLifecycleEvent;
  *     desc  :
  * </pre>
  */
-public abstract class BasePresenter<V extends BaseView> implements LifecycleObserver {
+public abstract class BasePresenter<V extends BaseView> extends Utils.ActivityLifecycleCallbacks {
 
     private V                     mView;
     private Map<Class, BaseModel> mModelMap = new HashMap<>();
@@ -29,6 +30,7 @@ public abstract class BasePresenter<V extends BaseView> implements LifecycleObse
     void bindView(V view) {
         this.mView = view;
         onAttachView();
+        ActivityUtils.addActivityLifecycleCallbacks(mView.getActivity(), this);
     }
 
     public V getView() {
@@ -54,9 +56,16 @@ public abstract class BasePresenter<V extends BaseView> implements LifecycleObse
         return null;
     }
 
-    @CallSuper
-    @OnLifecycleEvent(Lifecycle.Event.ON_DESTROY)
-    public void onDestroyPresenter() {
+    @Override
+    public void onLifecycleChanged(@NonNull Activity activity, Lifecycle.Event event) {
+        super.onLifecycleChanged(activity, event);
+        if (event == Lifecycle.Event.ON_DESTROY) {
+            destroyPresenter();
+        }
+        LogUtils.i("onLifecycleChanged: " + event);
+    }
+
+    private void destroyPresenter() {
         if (mView != null) {
             mView.mPresenterMap.remove(this.getClass());
             mView.onDestroyView();
@@ -67,11 +76,5 @@ public abstract class BasePresenter<V extends BaseView> implements LifecycleObse
             }
         }
         mModelMap.clear();
-        LogUtils.i("onDestroyPresenter");
-    }
-
-    @OnLifecycleEvent(Lifecycle.Event.ON_ANY)
-    public void onLifecycleChanged(LifecycleOwner owner, Lifecycle.Event event) {
-        LogUtils.i(event.toString());
     }
 }