+++ lib/utilcode/src/main/java/com/blankj/utilcode/util/LogUtils.java
@@ -594,7 +594,7 @@ public final class LogUtils {
 
     private static void input2File(final String filePath, final String input) {
         if (CONFIG.mFileWriter == null) {
-            UtilsBridge.writeFileFromString(filePath, input);
+            UtilsBridge.writeFileFromString(filePath, input, true);
         } else {
             CONFIG.mFileWriter.write(filePath, input);
         }
@@ -622,13 +622,12 @@ public final class LogUtils {
         private IFileWriter mFileWriter;
 
         private Config() {
-            mDefaultDir = Utils.getApp().getFilesDir() + FILE_SEP + "log" + FILE_SEP;
-            mFileWriter = new IFileWriter() {
-                @Override
-                public void write(String file, String content) {
-
-                }
-            };
+            if (UtilsBridge.isSDCardEnableByEnvironment()
+                    && Utils.getApp().getExternalFilesDir(null) != null)
+                mDefaultDir = Utils.getApp().getExternalFilesDir(null) + FILE_SEP + "log" + FILE_SEP;
+            else {
+                mDefaultDir = Utils.getApp().getFilesDir() + FILE_SEP + "log" + FILE_SEP;
+            }
         }
 
         public final Config setLogSwitch(final boolean logSwitch) {